7.4_draw house
 - draw house with OpenGL tools

This program use OpenGL to draw one house, the example
from The Art and Science of C. You know, this book is too old
to refer, especially the sample code in the chapter 7.

To solve this problem, I use OpenGL build a DrawTools. You
can use it to write code, do exercises or something.

If you use Xcode on Mac, just add OpenGL and GLUT, there is
a youtube video about how to do that:
https://www.youtube.com/watch?v=LPkwd7GoLv0

But if you use other IDE on Windows or Unix/Linux, sorry.
You need google it to find answer.